Приведение всех id к типу Long

+++ src/main/java/bsw/iron/barcode_server/entity/Conversion.java
@@ -15,7 +15,7 @@ public class Conversion {
 
     @Id
     @Column(name = "ID_PEREDEL")
-    private Integer idConversion;
+    private Long idConversion;
 
     @Column(name = "NAME_PEREDEL")
     private String nameConversion;
